Add Rake addon for document symbols
This commit adds a first-party Rake addon to provide document symbols
for the `namespace` and `task` DSL methods. Although there aren't any
other examples of first-party addons in the repository, it seemed to
make sense to model it this way, due to the fact it should only have
effect on a specific subset of files (`Rakefile` and `*.rake` files).

We could have extracted this to a separate gem, but given Rake is very
much a core part of Ruby—and the functionality added here so minimal—it
didn't make sense to do so.

We also could have _not_ modelled it as an addon, and instead just built
it into to `Listener::DocumentSymbol` directly. However, the `namespace`
and `task` DSL methods are specific to rake, so I wanted to limit the
files that these document symbols are generated for. Doing tat in an
addon seemed to make more sense than having conditionals inside the main
`DocumentSymbol` class.