Use set for addon_classes
This allows us to manually add a class to `addon_classes` during tests,
which might already exist if it hasn't been cleared yet. This isn't an
issue at the moment, because there aren't any real addons loaded in
tests. However, I'd like to add a first-party addon in an upcoming
commit which will need to be manually readded to `addon_classes` if I
want to test it.